UHYU-93 refactor: erd 설계에 따른 user 엔티티 수정 및 일부파일 위치 변경
